requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
PyJWT==2.3.0
pymongo==4.0.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
tinydb==4.7.0
twilio==7.7.1
urllib3==1.26.8
Werkzeug==2.0.3
